Utvider klassehierkiet, Lonnstilskudd som egen klasse for å ikke duplisere alt i VarigLonnstilskudd og MidlertidigLonnstilskudd
